+++ src/main/java/racing/Car.java
@@ -1,4 +1,4 @@
-package step3;
+package racing;
 
 import java.util.ArrayList;
 import java.util.List;
