+++ src/main/java/racing/GameMain.java
@@ -1,4 +1,4 @@
-package step3;
+package racing;
 
 import java.util.List;
