+++ src/main/java/racing/InputView.java
@@ -1,4 +1,4 @@
-package step3;
+package racing;
 
 import java.util.Scanner;
