+++ src/main/java/racing/RacingGame.java
@@ -1,7 +1,6 @@
-package step3;
+package racing;
 
 import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
