--- src/main/java/step2/InputView.java
@@ -1,18 +1,0 @@
-package step2;
-
-import java.util.Scanner;
-
-public class InputView {
-
-    public int getCarNum() {
-        System.out.println("자동차 대수는 몇 대 인가요?");
-        Scanner scanner = new Scanner(System.in);
-        return scanner.nextInt();
-    }
-
-    public int getTrialNum() {
-        System.out.println("시도할 회수는 몇 회 인가요?");
-        Scanner scanner = new Scanner(System.in);
-        return scanner.nextInt();
-    }
-}
